requirements.txt
# Django
Django==5.0.2
django-cors-headers==4.3.1
django-filter==23.5
django-environ==0.11.2

# Django REST Framework
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.27.1

# Database
psycopg2-binary==2.9.9
redis==5.0.1

# File storage
boto3==1.34.34
minio==7.2.3

# Async tasks
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.1

# AI/ML
anthropic==0.18.0
openai==1.7.0
Pillow==10.2.0

# Utilities
python-decouple==3.8
python-dateutil==2.8.2
pytz==2023.3.post1

# Development
ipython==8.18.1
django-extensions==3.2.3
django-debug-toolbar==4.2.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
Faker==20.1.0